Final touches and adjustments
Fixed inventory bugs and added item and thumbail. (Use I to open inventory, P to interact with chests)
Added enemies to the dungeon and added chests to regain health also.
Created a stronger enemy at the end of the dungeon to be a boss with separate functionality.
Cleaned up the map to get rid of anything that we weren't going to use.